Optimize on `dev` instead of modifying `release`.
`opt-level=2` seemed like the lowest level that could run games at full
frame rate. (`release` uses `opt-level=3`.)

This configuration seems simpler and keeps `release` as the profile for
actual releases, i.e. not for debugging while we are developing.
Additionally, the main reason I made this commit is that `dev` defaults
to `incremental=true` and `release` does not.